#54 Refactor getting current account
#54 Adds the util directory and adds the UserAccessor class to directly access the current account and remove duplication
